requirements.txt
roboflow
tensorflow
keras
numpy
matplotlib
inference-sdk
streamlit
python-dotenv
opencv-python